refactor: eliminate old infoview; fix layout; insert default infoview
