+++ client/src/css/app.css
@@ -13,7 +13,6 @@ html {
   --ff-primary: Roboto;
 }
 
-
 /* General styling */
 
 body {
@@ -92,22 +91,8 @@ em {
 
 /* App Bar */
 
-#root {
-  height: 100%;
-}
-
-.app {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
 .app-bar {
-  flex: 0;
   background: var(--clr-primary);
-  display: flex;
-  position: relative;
-  flex-direction: row;
   justify-content: space-between;
   align-items: center;
   padding: 1.1em;
@@ -129,14 +114,6 @@ em {
   margin: 0;
   /* margin: 0 1em; */
 }
-
-.app-content {
-  height: 100%;
-  flex: 1;
-  min-height: 0;
-  display: flex;
-}
-
 
 .markdown li ul, .markdown li ol {
 	margin:0 1.5em;
