+++ client/src/css/layout.css
@@ -1,0 +1,92 @@
+html, body {
+  height: 100%;
+  margin: 0;
+}
+
+#root {
+  height: 100vh;
+  width: 100vw;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+}
+
+.app-bar {
+  flex-shrink: 0;
+  background: var(--clr-primary);
+  display: flex;
+  position: relative;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  padding: 1.1em;
+  filter: drop-shadow(0 0 5px rgba(0,0,0,0.5));
+  z-index: 20;
+}
+
+.app-content {
+  flex: 1;
+  min-height: 0;
+  position: relative;
+}
+
+.level, .welcome {
+  display: flex;
+  height: 100%;
+}
+
+.inventory-panel, .exercise-panel, .doc-panel, .introduction-panel, .chat-panel, .panel {
+  display: flex;
+  min-height: 0;
+  flex-direction: column;
+  position: relative;
+  overflow-y: hidden;
+}
+
+/* World Tree Panel */
+
+.world-tree {
+    flex: 1;
+    min-height: 0;
+    overflow-y: auto;
+}
+
+.world-selection {
+  max-width: 100%;
+  display: block;
+  margin: auto;
+}
+
+/* Exercise Panel */
+.exercise {
+  gap: 8px;
+}
+
+.codeview {
+  flex: 1;
+}
+
+.infoview {
+  flex: 1;
+  display: flex;
+  border: 1px solid green;
+  padding-top: 1em;
+  padding-bottom: 0;
+  padding-top: 0em;
+}
+
+.infoview iframe {
+  width: 100%;
+  height: 100%;
+}
+
+/* Inventory Panel */
+
+.tab-bar {
+  flex-shrink: 0;
+}
+
+.inventory-list {
+  flex: 1;
+  overflow-y: auto;
+}
