+++ client/src/css/typewriter.css
@@ -1,0 +1,9 @@
+
+
+.typewriter-canvas {
+  position: absolute;
+  /* width: 100%;
+  height: 100%; */
+  background-color: white;
+  border: 1px solid red;
+}
