+++ client/src/css/world_tree.css
@@ -44,14 +44,6 @@ svg .disabled {
   cursor: default;
 }
 
-.world-selection {
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 2em;
-  max-width: 100%;
-}
-
 .world-selection-menu {
   right: 1em;
   top: 1em;
